Moving Travis related methods to its own class.
